requirements.txt
# Databricks Core Dependencies
databricks-connect==17.1.2
databricks-sdk==0.65.0

# Claude Agent SDK
claude-agent-sdk>=0.1.0
anthropic>=0.18.0

# CLI Framework
typer[all]>=0.9.0
rich>=13.0.0

# LiteLLM for proxy support (Databricks FMAPI)
litellm>=1.40.0

# MLflow for observability
mlflow>=3.5.0

# Async HTTP client
aiohttp>=3.9.0

# Note: Databricks CLI v2 is a standalone binary, install separately:
# curl -fsSL https://raw.githubusercontent.com/databricks/setup-cli/main/install.sh | sh

# Environment and Configuration
python-dotenv==1.0.1

# API Framework (for backend)
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# MCP Server Dependencies
mcp>=1.12.4
httpx==0.28.1
pyyaml>=6.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Development Tools
black==24.10.0
ruff==0.9.1
mypy==1.14.1
ipython==8.31.0

# Data Processing
pandas==2.2.3
numpy==1.26.4
pyarrow==18.1.0

# Logging and Monitoring
structlog==24.4.0
rich==13.9.4